kadai lesson 15

+++ app/Http/Controllers/TasksController.php
@@ -4,31 +4,31 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Task;    // 追加
+use App\Task;
 
 class tasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    // getでtasks/にアクセスされた場合の「一覧表示処理」
+
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy( 'created_at', 'desc')->paginate(10);
+            
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
-    }
+            ];
+            return view('tasks.index', $data);
+        }
+        return view('welcome');
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
+    }   
+
+
     public function create()
     {
         $task = new Task;
@@ -37,34 +37,30 @@ class tasksController extends Controller
             'task' => $task,
         ]);
     }
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    // postでtasks/にアクセスされた場合の「新規登録処理」
+
+
    public function store(Request $request)
     {
         $this->validate($request, [
-            'status' => 'required|max:10',   // 追加-
+            'status' =>  'required|max:10',   // 追加-
             'content' => 'required|max:191',
         ]);
 
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+             
+        ]);
+        
         $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        $task ->status = $request->status;    // 追加
+        $task ->content = $request->content;
 
         return redirect('/');
     }    
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    // getでtasks/idにアクセスされた場合の「取得表示処理」
+
+
+
     public function show($id)
     {
         $task = Task::find($id);
@@ -73,13 +69,8 @@ class tasksController extends Controller
             'task' => $task,
         ]);
     }
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-     // getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
+
+
     public function edit($id)
     {
         $task = Task::find($id);
@@ -88,14 +79,8 @@ class tasksController extends Controller
             'task' => $task,
         ]);
     }
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    // putまたはpatchでmessages/idにアクセスされた場合の「更新処理」
+
+
     public function update(Request $request, $id)
     {
         $this->validate($request, [
@@ -110,18 +95,16 @@ class tasksController extends Controller
 
         return redirect('/');
     }
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    // deleteでmessages/idにアクセスされた場合の「削除処理」
+
+
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
-        return redirect('/');
+        return redirect();
     }
 }
